# public/openapi/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DebtResolve API",
        "contact": {
            "email": "jadiael@hotmail.com.br"
        },
        "license": {
            "name": "CC-BY-NC",
            "url": "https://creativecommons.org/licenses/by-nc/4.0/"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/charges": {
            "get": {
                "tags": [
                    "Charges"
                ],
                "summary": "List all charges",
                "description": "Returns a list of all charges",
                "operationId": "e5a56f95598e083b376960fc46152c08",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Message about the operation status",
                                            "type": "string",
                                            "example": "Charge listing completed successfully"
                                        },
                                        "data": {
                                            "description": "Data containing the list of charges",
                                            "properties": {
                                                "charges": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Charge"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Charges"
                ],
                "summary": "Create a new charge",
                "description": "Creates a new charge and related installments",
                "operationId": "c5bdf89b8c07cf1688a25dcead0550c8",
                "requestBody": {
                    "description": "Charge data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "description",
                                    "amount",
                                    "installments_number",
                                    "due_day",
                                    "collector_id",
                                    "debtor_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the charge",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Description of the charge",
                                        "type": "string"
                                    },
                                    "amount": {
                                        "description": "Total debt amount",
                                        "type": "integer"
                                    },
                                    "installments_number": {
                                        "description": "Number of installments",
                                        "type": "integer"
                                    },
                                    "due_day": {
                                        "description": "Due day",
                                        "type": "integer"
                                    },
                                    "collector_id": {
                                        "description": "Collector user ID",
                                        "type": "integer"
                                    },
                                    "debtor_id": {
                                        "description": "Debtor user ID",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Charge successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Message about the operation status",
                                            "type": "string",
                                            "example": "Charge successfully created"
                                        },
                                        "data": {
                                            "description": "Data containing the created charge and its installments",
                                            "properties": {
                                                "charge": {
                                                    "$ref": "#/components/schemas/Charge"
                                                },
                                                "installments": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Installment"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Message about the error",
                                            "type": "string",
                                            "example": "You need to choose whether you are the collector or the debtor of the charge"
                                        },
                                        "errors": {
                                            "description": "Detailed error messages",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict - Charge already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Message about the error",
                                            "type": "string",
                                            "example": "This charge already exists"
                                        },
                                        "errors": {
                                            "description": "Detailed error messages",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unexpected error when creating resource",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Message about the error",
                                            "type": "string",
                                            "example": "Unexpected error when creating resource"
                                        },
                                        "errors": {
                                            "description": "Detailed error messages",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/charges/{charge}": {
            "get": {
                "tags": [
                    "Charges"
                ],
                "summary": "Get a specific charge",
                "description": "Returns a specific charge by ID",
                "operationId": "602ce8a5af8f810ed36716228dd3a478",
                "parameters": [
                    {
                        "name": "charge",
                        "in": "path",
                        "description": "ID of the charge to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "Message about the operation status",
                                            "type": "string",
                                            "example": "Charge found successfully"
                                        },
                                        "data": {
                                            "description": "Data containing the specific charge",
                                            "properties": {
                                                "charge": {
                                                    "$ref": "#/components/schemas/Charge"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Charge not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Message about the error",
                                            "type": "string",
                                            "example": "Charge not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Charge": {
                "properties": {
                    "id": {
                        "description": "Unique identifier for the charge",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the charge",
                        "type": "string",
                        "example": "Monthly rent"
                    },
                    "description": {
                        "description": "Description of the charge",
                        "type": "string",
                        "example": "Rent payment for the month"
                    },
                    "amount": {
                        "description": "Total amount of the charge",
                        "type": "number",
                        "example": 500
                    },
                    "installments_number": {
                        "description": "Number of installments for the charge",
                        "type": "integer",
                        "example": 12
                    },
                    "due_day": {
                        "description": "Due day for payment",
                        "type": "integer",
                        "example": 15
                    },
                    "collector_id": {
                        "description": "ID of the user who is the collector for this charge",
                        "type": "integer",
                        "example": 1
                    },
                    "debtor_id": {
                        "description": "ID of the user who is the debtor for this charge",
                        "type": "integer",
                        "example": 2
                    },
                    "collector": {
                        "$ref": "#/components/schemas/User"
                    },
                    "debtor": {
                        "$ref": "#/components/schemas/User"
                    },
                    "installments": {
                        "description": "List of installments associated with this charge",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Installment"
                        }
                    }
                },
                "type": "object"
            },
...
            "Installment": {
                "properties": {
                    "id": {
                        "description": "Unique identifier for the installment",
                        "type": "integer",
                        "example": 1
                    },
                    "value": {
                        "description": "Value of the installment",
                        "type": "number",
                        "example": 100
                    },
                    "installment_number": {
                        "description": "Number of the installment",
                        "type": "integer",
                        "example": 1
                    },
                    "due_date": {
                        "description": "Due date for payment",
                        "type": "string",
                        "format": "date",
                        "example": "2023-11-30"
                    },
                    "paid": {
                        "description": "Indicates if the installment is paid or not",
                        "type": "boolean",
                        "example": true
                    },
                    "user_id": {
                        "description": "ID of the user associated with the installment",
                        "type": "integer",
                        "example": 1
                    },
                    "awaiting_approval": {
                        "description": "Indicates if the installment is awaiting approval",
                        "type": "boolean",
                        "example": false
                    },
                    "payment_proof": {
                        "description": "URL or path to the payment proof file",
                        "type": "string",
                        "example": "https://example.com/payment-proof.jpg"
                    },
                    "charge": {
                        "$ref": "#/components/schemas/Charge"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "User's name",
                        "type": "string"
                    },
                    "email": {
                        "description": "User's email address",
                        "type": "string",
                        "format": "email"
                    },
                    "email_verified_at": {
                        "description": "Date/time when email was verified",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "Date/time of user creation",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Date/time of user update",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "Bearer"
            }
        }
    },
...
}